requirements.txt
fastapi
uvicorn
python-docx
pydantic
streamlit
requests
groq  # or use openai if using OpenAI-compatible client